requirements.txt
pytest==9.0.2
pytest-playwright==0.7.2
allure-pytest==2.13.5
requests==2.32.3
faker==37.1.0
pyyaml==6.0.2
